+++ changelog.md
@@ -1,4 +1,8 @@
 # Changelog
+## v0.2.4 (18-04-2026)
+- adding user_id di tb user_responses
+- create migrations user_id
+
 ## v0.2.3 (18-04-2026)
 - bebenah code Vue (pemisahan js dan css) all page